+++ src/main/java/command/AddDeadlineCommand.java
@@ -1,24 +1,27 @@
+package command;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.temporal.ChronoUnit;
+import task.*;
+import duke.*;
 
-public class AddEventCommand extends AddCommand {
+public class AddDeadlineCommand extends AddCommand {
 
     private String desc;
-    private LocalDate at;
+    private LocalDate by;
 
-    public AddEventCommand(String desc, String at) {
+    public AddDeadlineCommand(String desc, String by) {
         this.desc = desc;
-        this.at = LocalDate.parse(at);
+        this.by = LocalDate.parse(by);
     }
 
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
         // tasks
-        Event newEvent = new Event(desc, at);
-        tasks.add(newEvent);
+        Deadline newDeadline = new Deadline(desc, by);
+        tasks.add(newDeadline);
         // ui
-        String response = respond(newEvent, tasks.size());
+        String response = respond(newDeadline, tasks.size());
         ui.showResponse(response);
         // storage
         storage.save(tasks);
