+++ src/main/java/command/Command.java
@@ -1,0 +1,9 @@
+package command;
+import duke.*;
+import task.*;
+public abstract class Command {
+
+    public abstract void execute(TaskList tasks, Ui ui, Storage storage);
+
+    public abstract boolean isExit();
+}
