+++ src/main/java/duke_exception/DukeDeleteException.java
@@ -1,3 +1,4 @@
+package duke_exception;
 public class DukeDeleteException extends DukeException {
     public DukeDeleteException() {
         super("OOPS!!! The index of a delete cannot be empty.");
