+++ src/main/java/duke_exception/DukeDoneException.java
@@ -1,3 +1,4 @@
+package duke_exception;
 public class DukeDoneException extends DukeException {
     public DukeDoneException() {
         super("OOPS!!! The index of a done cannot be empty.");
