+++ src/main/java/duke_exception/DukeEventException.java
@@ -1,3 +1,4 @@
+package duke_exception;
 public class DukeEventException extends DukeException {
     public DukeEventException() {
         super("OOPS!!! The description of a event cannot be empty.");
