+++ src/main/java/duke_exception/DukeException.java
@@ -1,3 +1,4 @@
+package duke_exception;
 public class DukeException extends Exception {
     public DukeException(String message) {
         super(message);
