+++ src/main/java/duke_exception/DukeTodoException.java
@@ -1,3 +1,4 @@
+package duke_exception;
 public class DukeTodoException extends DukeException {
     public DukeTodoException() {
         super("OOPS!!! The description of a todo cannot be empty.");
